LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/svg-root-size-container-expected.txt

Green
Green

FAIL SVG text querying SVG root size container assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL div in foreignObject querying SVG root size container assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
